requirements.txt
fastapi~=0.110.0
uvicorn[standard]~=0.23.2
fastapi-utilities==0.2.0
pydantic~=2.6.3
psycopg2-binary~=2.9.9
python-dotenv~=1.0.1
passlib~=1.7.4
python-jose~=3.0.0
python-multipart==0.0.9
bcrypt==4.1.2
starlette~=0.36.3
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
pydantic[email]

pymongo~=4.9.2
PyJWT~=2.9.0
bson~=0.5.10
motor~=3.6.0
dnspython~=2.3.0
idna~=3.8
cryptography~=42.0.5
future~=0.18.3
anyio~=4.0.0
h11~=0.14.0
jupyter~=1.0.0
pip~=24.0
attrs~=23.1.0
Pillow~=10.0.0
tornado~=6.3.3
Jinja2~=3.1.2
rsa~=4.9
pyasn1~=0.5.1
six~=1.16.0
ecdsa~=0.18.0
pytz~=2023.3
sniffio~=1.3.0
setuptools~=65.5.0
overrides~=7.4.0
cffi~=1.15.1
pyinstaller~=5.10.1
click~=8.1.7
ipython~=8.14.0
SQLAlchemy~=2.0.30
lxml~=5.3.0
matplotlib~=3.7.2
MarkupSafe~=2.1.3
scipy~=1.11.3
python-dateutil~=2.8.2
requests~=2.31.0
websockets~=12.0
certifi~=2024.8.30
httptools~=0.6.1
watchfiles~=0.21.0
croniter~=1.4.1
greenlet~=3.0.3
psutil~=5.9.4
defusedxml~=0.7.1
decorator~=4.4.2
PyYAML~=6.0.1
packaging~=23.1
pyparsing~=3.0.9